Bug#13945257 NEW COMPILATION ERRORS/WARNINGS WITH XCODE 4.3.2
Fix various new compiler warnings and one compiler error
reported by Clang 3.1 / XCode 4.3.2 of Mac OSX.

See per-file messages for details.

+++ sql/log_event.cc
@@ -1316,7 +1316,8 @@ Log_event* Log_event::read_log_event(IO_CACHE* file,
   Log_event *res=  0;
 #ifndef max_allowed_packet
   THD *thd=current_thd;
-  uint max_allowed_packet= thd ? thd->variables.max_allowed_packet : ~(ulong)0;
+  ulong max_allowed_packet= thd ? thd->variables.max_allowed_packet :
+                                  ~(ulong)0;
 #endif
 
   ulong const max_size=
